refactor(agent-tars-web-ui): impl
